+++ packages/config/eslint-preset.js
@@ -42,6 +42,7 @@ module.exports = {
         "packages/spor-badge-react-native/",
         "packages/spor-accordion-react-native/",
         "packages/spor-bottom-drawer-react-native/",
+        "packages/spor-alert-react-native/",
       ],
     },
   },