+++ packages/spor-react-native/src/index.tsx
@@ -1,6 +1,7 @@
 export * from "@vygruppen/spor-button-react-native";
 export * from "@vygruppen/spor-card-react-native";
 export * from "@vygruppen/spor-i18n-react";
+export * from "@vygruppen/spor-alert-react-native";
 export * from "@vygruppen/spor-layout-react-native";
 export * from "@vygruppen/spor-loader-react-native";
 export * from "@vygruppen/spor-provider-react-native";