+++ packages/spor-theme-react-native/src/components/badge.tsx
@@ -41,4 +41,8 @@ export const badgeColorSchemes = {
     backgroundColor: "lightRed",
     borderColor: "brightRed",
   },
+  "night-train": {
+    backgroundColor: "darkBlue",
+    borderColor: "darkGrey",
+  },
 };