+++ packages/spor-badge-react-native/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @vygruppen/spor-badge-react-native
 
+## 1.0.5
+
+### Patch Changes
+
+- 9d2fba0: Add color scheme for night-train badge
+- Updated dependencies [9d2fba0]
+  - @vygruppen/spor-theme-react-native@0.2.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [c0b0f01]
+  - @vygruppen/spor-theme-react-native@0.2.4
+
 ## 1.0.3
 
 ### Patch Changes