+++ packages/spor-card-react-native/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vygruppen/spor-card-react-native
 
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [9d2fba0]
+  - @vygruppen/spor-theme-react-native@0.2.5
+  - @vygruppen/spor-button-react-native@0.2.5
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [c0b0f01]
+  - @vygruppen/spor-theme-react-native@0.2.4
+  - @vygruppen/spor-button-react-native@0.2.4
+
 ## 0.0.12
 
 ### Patch Changes