+++ packages/spor-message-box-react-native/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vygruppen/spor-message-box-react-native
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [9d2fba0]
+  - @vygruppen/spor-theme-react-native@0.2.5
+  - @vygruppen/spor-button-react-native@0.2.5
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [c0b0f01]
+  - @vygruppen/spor-theme-react-native@0.2.4
+  - @vygruppen/spor-button-react-native@0.2.4
+
 ## 1.0.2
 
 ### Patch Changes