+++ packages/spor-provider-react-native/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vygruppen/spor-provider-react-native
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [9d2fba0]
+  - @vygruppen/spor-theme-react-native@0.2.5
+
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [c0b0f01]
+  - @vygruppen/spor-theme-react-native@0.2.4
+
 ## 0.1.12
 
 ### Patch Changes