+++ packages/spor-theme-react-native/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vygruppen/spor-theme-react-native
 
+## 0.2.5
+
+### Patch Changes
+
+- 9d2fba0: Add color scheme for night-train badge
+
+## 0.2.4
+
+### Patch Changes
+
+- c0b0f01: Lagt til første versjon av alert komponent
+
 ## 0.2.3
 
 ### Patch Changes